// apps/frontend/src/enums/RoutePath.ts
export enum RoutePath {
  HOME = "/",
  USER = "/user",
  DASHBOARD = "/dashboard",
  LOGIN = "/login",
  OFFER_EDITOR = "/offer/edit/:id?",
  OFFER_MY = "/offer/my/",
  OFFER_MY_ARCHIVE = "/offer/my/archive",
  OFFER_ARCHIVE = "/offer/archive",
  OFFER_PREVIEW = "/offer/:id",
  DICTIONARIES = "/dict/",
}